finished

+++ clash.c
@@ -20,7 +20,7 @@ int
 toUpperString(char *, int);
 
 /*
- *
+ * 用于刷新clash
  */
 int 
 main(int argc, char **argv)
@@ -40,78 +40,105 @@ main(int argc, char **argv)
 }
 
 /*
- *
+ * 根据processname通过查找进程号
  */
 int 
 findProcess(const char * const processname)
 {
-    char buff[BUFSIZ];
+    char command[BUFSIZ];
     FILE * fp;
     pid_t pp;
 
-    memset(buff, 0, sizeof(buff));
+    memset(command, 0, sizeof(command));
     
-    sprintf(buff, "pgrep %s", processname);
-    fp=popen(buff, "r");
+    /*
+     * 使用pgrep通过shell查找
+     */
+    sprintf(command, "pgrep %s", processname);
+    fp=popen(command, "r");
     if(fp==NULL)
     {
         perror("popen");
         return -1;
     }
+    fscanf(fp, "%d", &pp);
 
-    pp=fscanf(fp, "%d", &pp);
-    printf("kill %s: %d\n", processname, pp);
-    kill(pp, SIGKILL);
+    /*
+     * 若之前存在clash进程则杀死之前的clash进程
+     */
+    if(pp!=0)
+    {
+        printf("LOG: kill %s: %d\n", processname, pp);
+        kill(pp, SIGKILL);
+    }
     pclose(fp);
 
+    /*
+     * 重新启动clash
+     */
     restartProcess(processname);
 
     return 0;
 }
 
 /*
- * 
+ * 根据processname, 重新启动进程
  */
 int 
 restartProcess(const char * const processname)
 {
-    char buff[BUFSIZ];
+    char command[BUFSIZ];
     char pn[strlen(processname)+1];
     pid_t pp;
     FILE * fp;
 
+    /*
+     * 根据processname获取环境变量${CLASH}
+     */
     memset(pn,0,sizeof(pn));
+    strncpy(pn, processname, strlen(processname));
     if(toUpperString(pn, strlen(pn))<0)
     {
         perror("lower2Upper");
         return -1;
     }
 
-    sprintf(buff, "${%s}", pn);
-    fp = popen(buff, "r");
+    /*
+     * 根据${CLASH}启动clash进程
+     */
+    sprintf(command, "${%s}", pn);
+    fp = popen(command, "r");
     if(fp==NULL)
     {
         perror("popen");
         return -1;
     }
 
-    sprintf(buff, "pgrep %s", processname);
-    fp=popen(buff, "r");
+    /*
+     * 输出调试信息测试是否成功启动
+     */
+    sprintf(command, "pgrep %s", processname);
+    fp=popen(command, "r");
     if(fp==NULL)
     {
         perror("popen");
         return -1;
     }
-    
-    pp=fscanf(fp, "%d", &pp);
-    printf("restart %s: %d\n", pn, pp);
+    fscanf(fp, "%d", &pp);
+    if(pp!=0)
+    {
+        printf("restart %s: %d\n", pn, pp);
+    } else {
+        perror("popen");
+        return -1;
+    }
     pclose(fp);
 
     return 0;
 }
 
 /*
- *
+ * 将小写的origin字符串全部转变为大写
  */
 int 
 toUpperString(char * origin, int len)
